move MetaMixin class to meta_view_mixins.py
